feat: Split creator into tabbed, modular applications
This commit completes the planned refactoring of the course creator. The application is now split into three distinct, more secure pages, unified under a single tabbed interface.

This resolves the security concern of overly broad Content Security Policies and improves the maintainability of the codebase.

Key changes:
- The monolithic `course-creator.html` has been replaced by a tabbed "shell".
- Three new creator pages have been created: `cloud_creator.html`, `webllm_creator.html`, and `ollama_creator.html`, each with a specific and strict CSP.
- The JavaScript has been refactored into shared modules and three new entry points (`cloud_main.js`, `webllm_main.js`, `ollama_main.js`).
- A new `shell_main.js` handles the tab-switching logic.
- Bugs related to the initial refactoring have been fixed, ensuring all new pages are fully functional.
